Refactor: 회원가입 관련 코드 수정
