+++ lib/widget/register/email_input_field.dart
@@ -1,29 +1,30 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:yangjataekil/controller/register_controller.dart';
+import 'package:yangjataekil/theme/app_color.dart';
 
-class EmailInputField extends StatelessWidget {
+class EmailInputField extends GetView<RegisterController> {
   const EmailInputField({
     Key? key,
     required this.hintText,
     required this.obscureText,
-    required this.controller,
+    required this.textController,
     required this.icon,
   }) : super(key: key);
 
   final String hintText;
   final bool obscureText;
-  final TextEditingController controller;
+  final TextEditingController textController;
   final Icon icon;
 
   @override
   Widget build(BuildContext context) {
-    final RegisterController controller = Get.find();
+
     return TextFormField(
       onChanged: (value) {
         controller.updateEmail(value);
       },
-      controller: this.controller,
+      controller: textController,
       obscureText: obscureText,
       decoration: InputDecoration(
         hintText: hintText,
@@ -33,7 +34,7 @@ class EmailInputField extends StatelessWidget {
         ),
         focusedBorder: OutlineInputBorder(
           borderRadius: BorderRadius.circular(50),
-          borderSide: const BorderSide(color: Color(0xffFF9297), width: 0.5),
+          borderSide: const BorderSide(color: AppColors.btnBackgroundColor, width: 0.5),
         ),
         suffixIcon: Container(
           padding: EdgeInsets.zero,
